backend/requirements.txt
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
requests>=2.31.0
drf-spectacular>=0.27.1
